+++ src/Hacks/FullOptions.cpp
@@ -5,20 +5,24 @@
 
 using namespace geode::prelude;
 
-bool ig = false;
+namespace FullOptions
+{
+    bool ignore = false;
+};
+using namespace FullOptions;
 
 class $modify (OptionsLayerExt, OptionsLayer)
 {
     void onPauseOptions(CCObject* sender)
     {
-        ig = true;
+        ignore = true;
 
         if (auto pause = getChildOfType<PauseLayer>(CCScene::get(), 0))
         {
             pause->onSettings(sender);
         }
         
-        ig = false;
+        ignore = false;
     }
 
     void customSetup()
@@ -29,15 +33,13 @@ class $modify (OptionsLayerExt, OptionsLayer)
             return;
 
         auto btn = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_optionsBtn02_001.png"), this, menu_selector(OptionsLayerExt::onPauseOptions));
+        btn->setID("normal-settings-btn"_spr);
 
-        if (auto l = m_mainLayer)
+        if (auto m = getChildOfType<CCMenu>(m_mainLayer, -2))
         {
-            if (auto m = getChildOfType<CCMenu>(l, 0))
-            {
-                btn->setPosition(ccp(145, -185));
+            btn->setPosition(ccp(18, -50));
 
-                m->addChild(btn);
-            }
+            m->addChild(btn);
         }
     }
 };
@@ -46,7 +48,7 @@ class $modify (PauseLayer)
 {
     void onSettings(cocos2d::CCObject* sender)
     {
-        if (Client::GetModuleEnabled("full-options") && !ig)
+        if (Client::GetModuleEnabled("full-options") && !ignore)
         {
             auto layer = OptionsLayer::create();
             CCDirector::sharedDirector()->getRunningScene()->addChild(layer, 69);